docker/requirements.txt
google-auth==2.27.0
google-api-python-client==2.108.0
pandas==2.0.3
boto3==1.34.20
fastparquet==2023.4.0
numpy==1.24.3
